feat: add first number task, restructure project
