+++ strings_tasks/lib/numbers_tasks.dart
@@ -1,0 +1,30 @@
+/* *******************************************************************************************
+ *                                                                                           *
+ * Please read the following docs:                                                           *
+ * https://dart.dev/guides/language/numbers                                                  *
+ * https://dart.dev/guides/language/language-tour#numbers-and-booleans                       *
+ *                                                                                           *
+ * https://api.dart.dev/stable/2.19.1/dart-core/int-class.html                               *
+ * https://api.dart.dev/stable/2.19.1/dart-core/double-class.html                            *
+ * https://api.dart.dev/stable/2.19.1/dart-core/BigInt-class.html                            *
+ * https://api.dart.dev/stable/2.19.1/dart-core/num-class.html                               *
+ *                                                                                           *
+ * and some js tutorials before implementing tasks:                                          *
+ * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Numbers_and_dates           *
+ * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number   *
+ * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math     *
+ *                                                                                           *
+ ******************************************************************************************* */
+
+/// Returns an area of a rectangle given by width and height.
+///
+/// @param {number} width
+/// @param {number} height
+/// @return {number}
+///
+/// @example:
+///   5, 10 => 50
+///   5, 5  => 25
+int? getRectangleArea(width, height) {
+  return width * height;
+}
