+++ strings_tasks/lib/strings_tasks.dart
@@ -316,8 +316,8 @@ bool? isString(dynamic value) {
 ///   'K♠' => 51
 
 int? getCardId(String x) {
-  // const playid = 'K♠';
-  final cards = [
+  // const x = 'K♠';
+  const cards = [
     'A♣',
     '2♣',
     '3♣',
@@ -375,65 +375,3 @@ int? getCardId(String x) {
   final cardIndex = cards.indexOf(x);
   return cardIndex != -1 ? cardIndex : null;
 }
-
-
-
-
-// int? getCardId(x) {
-//   // const x = '7♣';
-//   const array = [
-//     'A♣',
-//     '2♣',
-//     '3♣',
-//     '4♣',
-//     '5♣',
-//     '6♣',
-//     '7♣',
-//     '8♣',
-//     '9♣',
-//     '10♣',
-//     'J♣',
-//     'Q♣',
-//     'K♣',
-//     'A♦',
-//     '2♦',
-//     '3♦',
-//     '4♦',
-//     '5♦',
-//     '6♦',
-//     '7♦',
-//     '8♦',
-//     '9♦',
-//     '10♦',
-//     'J♦',
-//     'Q♦',
-//     'K♦',
-//     'A♥',
-//     '2♥',
-//     '3♥',
-//     '4♥',
-//     '5♥',
-//     '6♥',
-//     '7♥',
-//     '8♥',
-//     '9♥',
-//     '10♥',
-//     'J♥',
-//     'Q♥',
-//     'K♥',
-//     'A♠',
-//     '2♠',
-//     '3♠',
-//     '4♠',
-//     '5♠',
-//     '6♠',
-//     '7♠',
-//     '8♠',
-//     '9♠',
-//     '10♠',
-//     'J♠',
-//     'Q♠',
-//     'K♠',
-//   ];
-//   return array.indexOf(x);
-// }
