+++ strings_tasks/test/core_dart_101_test.dart
@@ -1,0 +1,7 @@
+import 'numbers_tests.dart';
+import 'string_tests.dart';
+
+void main() {
+  stringTests();
+  numbersTests();
+}
